requirements.txt
duckdb>=0.10.0
polars>=0.20.0
pyyaml>=6.0
streamlit>=1.30.0
openpyxl>=3.1.0
pytest>=8.0.0
